requirements.txt
ultralytics
opencv-python
pillow
streamlit
matplotlib
seaborn
pyyaml

